Patterns::add method + refactor Pattern and HtmlCompressor
