+++ src/HtmlCompressor.php
@@ -2,23 +2,34 @@
 
 namespace WyriHaximus\HtmlCompress;
 
-use WyriHaximus\Compress\CompressorInterface;
-use WyriHaximus\HtmlCompress\Compressor\HtmlCompressor as DefaultCompressor;
+use voku\helper\HtmlMin as DefaultCompressor;
 
 final class HtmlCompressor implements HtmlCompressorInterface
 {
-    /**
-     * @var CompressorInterface
-     */
-    private $defaultCompressor;
+	
+	/** @var DefaultCompressor */
+	private $defaultCompressor;
+	
+	/** @var Patterns */
+	private $patterns;
 
     public function __construct(Patterns $patterns)
     {
-        $this->defaultCompressor = new DefaultCompressor($patterns);
+		$this->defaultCompressor = new DefaultCompressor();
+		$this->patterns = $patterns;
+        $this->compressor()->attachObserverToTheDomLoop($this->patterns); // Patterns $patters IS already a voku\helper\HtmlMinDomObserverInterface;
     }
 
     public function compress(string $html): string
     {
-        return $this->defaultCompressor->compress($html);
-    }
+		return $this->compressor()->minify($html);
+	}
+	
+	public function compressor(): DefaultCompressor{
+		return $this->defaultCompressor;
+	}
+	
+	public function patterns(): Patterns{
+		return $this->patterns;
+	}
 }
