--- src/HtmlMinObserver.php
@@ -1,40 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace WyriHaximus\HtmlCompress;
-
-use voku\helper\HtmlMin;
-use voku\helper\HtmlMinDomObserverInterface;
-use voku\helper\HtmlMinInterface;
-use voku\helper\SimpleHtmlDomInterface;
-
-final class HtmlMinObserver implements HtmlMinDomObserverInterface
-{
-    /** @var Patterns */
-    private $patterns;
-
-    public function __construct(Patterns $patterns)
-    {
-        $this->patterns = $patterns;
-    }
-
-    /**
-     * Receive dom elements before the minification.
-     *
-     * @param SimpleHtmlDomInterface $element
-     * @param HtmlMinInterface       $htmlMin
-     */
-    public function domElementBeforeMinification(SimpleHtmlDomInterface $element, HtmlMinInterface $htmlMin): void
-    {
-        $this->patterns->compress($element);
-    }
-
-    /**
-     * Receive dom elements after the minification.
-     *
-     * @param SimpleHtmlDomInterface $element
-     * @param HtmlMinInterface       $htmlMin
-     */
-    public function domElementAfterMinification(SimpleHtmlDomInterface $element, HtmlMinInterface $htmlMin): void
-    {
-    }
-}
