+++ src/Patterns.php
@@ -2,12 +2,18 @@
 
 namespace WyriHaximus\HtmlCompress;
 
-use voku\helper\SimpleHtmlDomInterface;
 
-final class Patterns
+use voku\helper\HtmlMinDomObserverInterface;
+use voku\helper\SimpleHtmlDomInterface;
+use voku\helper\HtmlMinInterface;
+
+final class Patterns implements HtmlMinDomObserverInterface
 {
     /** @var PatternInterface[] */
     private $patterns = [];
+	
+	/** @var HtmlMinInterface */
+    private $htmlMin;
 
     /**
      * @param PatternInterface ...$patterns
@@ -15,7 +21,15 @@ final class Patterns
     public function __construct(PatternInterface ...$patterns)
     {
         $this->patterns = $patterns;
-    }
+	}
+	
+
+	/**
+     * @param PatternInterface $pattern
+     */
+	public function add(PatternInterface $pattern){
+		$this->patterns[] = $pattern;
+	}
 
     public function compress(SimpleHtmlDomInterface $element): void
     {
@@ -26,5 +40,30 @@ final class Patterns
                 return;
             }
         }
+	}
+	
+	/**
+     * Receive dom elements before the minification.
+     *
+     * @param SimpleHtmlDomInterface $element
+     * @param HtmlMinInterface       $htmlMin
+     */
+	public function domElementBeforeMinification(SimpleHtmlDomInterface $element, HtmlMinInterface $htmlMin): void
+    {
+		$this->htmlMin = $htmlMin;
+
+        $this->compress($element);
     }
+
+    /**
+     * Receive dom elements after the minification.
+     *
+     * @param SimpleHtmlDomInterface $element
+     * @param HtmlMinInterface       $htmlMin
+     */
+    public function domElementAfterMinification(SimpleHtmlDomInterface $element, HtmlMinInterface $htmlMin): void
+    {
+	}
+	
+	
 }
